pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "qdata-adapter-kd-galaxy"
version = "0.1.0"
description = "QDataV2 adapter for kd-galaxy"
readme = "README.md"
license = {text = "AGPL-3.0"}
authors = [
    {name = "广东轻亿云软件科技有限公司", email = "opensource@qeasy.cloud"}
]
maintainers = [
    {name = "QData Team", email = "opensource@qeasy.cloud"}
]
requires-python = ">=3.11"
keywords = ["qdata", "adapter", "kd-galaxy", "integration", "qeasy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "qdata-adapter-base>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.28.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "ruff>=0.1.0",
    "bandit>=1.7.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]
all = [
    "qdata-adapter-kd-galaxy[dev,docs]"
]

[project.urls]
Homepage = "https://www.qeasy.cloud"
Documentation = "https://github.com/qeasy/qdata-adapter-kd-galaxy#readme"
Repository = "https://github.com/qeasy/qdata-adapter-kd-galaxy"
"Bug Tracker" = "https://github.com/qeasy/qdata-adapter-kd-galaxy/issues"
Changelog = "https://github.com/qeasy/qdata-adapter-kd-galaxy/blob/main/CHANGELOG.md"

# Entry points for adapter discovery
[project.entry-points."qdata.adapters"]
kd_galaxy = "qdata_adapter_kd_galaxy:KdGalaxyAdapter"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

# ========== Black Configuration ==========
[tool.black]
line-length = 100
target-version = ['py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# ========== isort Configuration ==========
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["qdata_adapter_kd_galaxy"]
known_third_party = ["pytest", "qdata_adapter", "httpx", "pydantic"]

# ========== MyPy Configuration ==========
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ========== Pytest Configuration ==========
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=qdata_adapter_kd_galaxy",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--strict-markers",
    "--strict-config",
]
minversion = "7.0"

# ========== Coverage Configuration ==========
[tool.coverage.run]
source = ["src/qdata_adapter_kd_galaxy"]
omit = [
    "*/tests/*",
    "*/examples/*",
    "*/docs/*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

# ========== Ruff Configuration ==========
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "ANN",
    "S",
    "TCH",
    "PTH",
    "COM",
    "C90",
]
ignore = [
    "ANN101",
    "ANN102",
    "COM812",
    "C901",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S101"]
"examples/*" = ["ANN"]

[tool.ruff.lint.isort]
known-first-party = ["qdata_adapter_kd_galaxy"]
known-third-party = ["pytest", "qdata_adapter", "httpx", "pydantic"]

[tool.ruff.lint.mccabe]
max-complexity = 15